#1: Use ChangeSets to group merged changes for manual processing.
When an error/conflict is encountered during merging a ChangeSet is
created and initialized to match the contents of the merged LogEntry.
The ChangeSet's title reflects the merged revision and its comment is
set to the rewritten message.

In order to implement this functionality, the no-commit setting has been
removed from the Subcherry Merge view's toolbar and is not part of the
merge wizard.

The merge wizard's summary page has been changed to allow users to
select the message rewrite as well as commit modes. Additionally,
instead of displaying total numbers, an overview of selected tickets and
revisions is displayed.